SSE/github/requirements.txt
fastmcp
GitPython
PyGithub
python-dotenv